pyproject.toml
[project]
name = "blog-post-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.12",
    "fastapi>=0.116.1",
    "fastembed>=0.7.1",
    "google-generativeai>=0.8.5",
    "googlesearch-python>=1.3.0",
    "langsmith>=0.4.15",
    "openai>=1.100.2",
    "openinference-instrumentation-agno>=0.1.13",
    "openrouter>=1.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "pycountry>=24.6.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.15.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.1",
    "tavily-python>=0.7.11",
    "uvicorn>=0.35.0",
    "watchdog>=6.0.0",
]
